refactor: rewrite imports for relocated components and utils
