+++ ui/src/App.js
@@ -3,9 +3,9 @@ import MainContainer from './pages/MainContainer';
 import { store, persistor } from './store';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
-import NavBar from './pages/NavigationBar/NavBar';
+import NavBar from 'components/NavigationBar/NavBar';
 import { Flex } from 'components';
-import MobileNavBar from './pages/NavigationBar/MobileNavBar';
+import MobileNavBar from 'components/NavigationBar/MobileNavBar';
 
 const App = () => {
   return (
