+++ ui/src/pages/MainContainer.js
@@ -10,21 +10,21 @@ import VisitUsPage from './discover/PlanAVisit';
 import DiscoverPage from './discover/DiscoverPage';
 import BBPage from './building-block/BBPage';
 import ShinePage from './shine/ShinePage';
-import PrivateRoute from './helpers/PrivateRoute';
+import PrivateRoute from './PrivateRoute';
 import Signup from './auth/Signup';
 import HomeContainer from './home/HomeContainer';
-import Footer from './footer/Footer';
+import Footer from 'components/Footer';
 import ConfirmEmailPage from './email/ConfirmEmailPage';
 import UserProfileContainer from './userProfile/UserProfileContainer';
 import CompleteUserProfileContainer from './userProfile/CompleteUserProfile';
 import GivingPage from './giving/GivingPage';
-import ClearCache from './helpers/ClearCache';
+import ClearCache from 'pages/screens/ClearCache';
 import RecoverPassword from './auth/RecoverPassword';
 import ResetPassword from './auth/ResetPassword';
 import PasswordSubmitted from './auth/PasswordSubmitted';
 import UserFormContainer from './forms/UserFormContainer';
 import ErrorPage from './screens/ErrorPage';
-import ScrollToTop from './helpers/ScrollToTop';
+import ScrollToTop from 'utils/ScrollToTop';
 import AboutUsContainer from './about/AboutUsContainer';
 import OnlineSermonContainer from './sermons/OnlineSermonContainer';
 import AdminHome from './admin/AdminHome';
