+++ ui/src/pages/admin/announcements/AdminAnnouncementContainer.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { customAxios as axios } from '../../helpers/customAxios';
+import { customAxios as axios } from 'utils/customAxios';
 import {
   Button,
   Heading,
