+++ ui/src/pages/admin/sermonNotes/AdminSermonNotesContainer.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback, useMemo } from 'react';
-import { customAxios as axios } from '../../helpers/customAxios';
+import { customAxios as axios } from 'utils/customAxios';
 import {
   Container,
   Heading,
