+++ ui/src/pages/auth/RecoverPassword.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { customAxios as axios } from '../helpers/customAxios';
+import { customAxios as axios } from 'utils/customAxios';
 import { useForm } from 'react-hook-form';
 import { ChevronLeftIcon } from '@chakra-ui/icons';
 import {
