+++ ui/src/pages/auth/ResetPassword.js
@@ -1,5 +1,5 @@
 import React, { useRef } from 'react';
-import { customAxios as axios } from '../helpers/customAxios';
+import { customAxios as axios } from 'utils/customAxios';
 import { useForm } from 'react-hook-form';
 import { ChevronLeftIcon } from '@chakra-ui/icons';
