+++ ui/src/pages/email/ConfirmEmailPage.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { customAxios as axios } from "../helpers/customAxios";
+import { customAxios as axios } from "utils/customAxios";
 import { Box, Container, Text } from "@chakra-ui/react";
 
 const ConfirmEmailPage = (props) => {
