+++ ui/src/pages/sermons/OnlineSermonContainer.js
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import { customAxios as axios } from '../helpers/customAxios';
+import { customAxios as axios } from 'utils/customAxios';
 import {
   HStack,
   VStack,
