+++ ui/src/pages/userProfile/UserProfileMobile.js
@@ -27,13 +27,13 @@ import {
 import { CheckCircleIcon } from '@chakra-ui/icons';
 import { useCallback, useEffect, useState } from 'react';
 import { useForm, Controller } from 'react-hook-form';
-import { customAxios as axios } from '../helpers/customAxios';
+import { customAxios as axios } from 'utils/customAxios';
 import {
   ministryTeamList,
   districtList,
   countryList,
   regionList,
-} from '../helpers/lists';
+} from 'utils/lists';
 import {
   settableDataFields,
   userDataCleanup,
@@ -41,7 +41,7 @@ import {
   updateUserDataRequest,
   getLoginOnlyFormsRequest,
   generatePublishedFormLinks,
-} from '../helpers/userInformationHelpers';
+} from 'utils/userInformationHelpers';
 import SermonNotesPagination from './SermonNotesPagination';
 
 const UserProfileMobile = (props) => {
